initial unit tests
