+++ .gitignore
@@ -149,3 +149,5 @@ data.json
 *.pkl
 # Local Netlify folder
 .netlify
+
+**/dist/
