+++ backend/.mocharc.json
@@ -1,0 +1,10 @@
+{
+  "loader": "ts-node/esm",
+  "extensions": ["ts", "tsx"],
+  "spec": [
+    "tests/**/*.spec.*"
+  ],
+  "watch-files": [
+    "src"
+  ]
+}
