+++ backend/package.json
@@ -7,7 +7,7 @@
     "start": "npx nodemon src/server.ts",
     "build": "npx tsc",
     "start:prod": "npm run build && node dist/server.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "mocha"
   },
   "author": "",
   "license": "ISC",
@@ -19,10 +19,18 @@
     "morgan": "^1.10.0"
   },
   "devDependencies": {
+    "@types/chai": "^4.3.4",
     "@types/express": "^4.17.14",
     "@types/express-session": "^1.17.5",
+    "@types/mocha": "^10.0.1",
     "@types/node": "^18.11.9",
+    "@types/sinon": "^10.0.13",
+    "@types/supertest": "^2.0.12",
+    "chai": "^4.3.7",
+    "mocha": "^10.2.0",
     "nodemon": "^2.0.20",
+    "sinon": "^15.0.1",
+    "supertest": "^6.3.3",
     "ts-node": "^10.9.1",
     "typescript": "^4.8.4"
   }
