+++ backend/src/server.ts
@@ -1,48 +1,14 @@
-import express from 'express';
-import bodyParser from "body-parser";
-import morgan from "morgan";
-import session from "express-session"
-import * as dotenv from 'dotenv';
-import "./typings" // Defines the types for session store data. Workaround with this bug: https://github.com/expressjs/session/issues/799
+import app from "./app"
 
-dotenv.config(); // Should load .env file directly
-
-import spotifyRoutes from "./routes/spotify";
-
-const app = express();
-
-app.use(bodyParser.json());
-app.use(bodyParser.urlencoded({
-  extended: true
-}));
-app.use(morgan("combined", {
-  skip: (req, res) => {
-    const url : string = req.url;
-
-    return url.includes("/static/")
-  }
-})); // Logger to listen to requests
-
-app.use(express.static("static/react"));
-
-app.get("/statusCheck", (req, res) => {
-  res.send("OK")
-});
-
-// Enable sessions for spotify authorization routes. This must be 
-// placed before the spotify routes
-app.set('trust proxy', 1);
-app.use(session({
-  secret: "This is extra secret here",
-  saveUninitialized: true,
-  resave: false,
-  cookie: {
-    maxAge: 60 * 60 * 1000 //In milliseconds, so this is one hour
-  }
-}));
-
-app.use('/api/spotify', spotifyRoutes)
-
+/**
+ * This snippet of code could be in the end of app.ts but it will 
+ * cause the unit tests to hang. By importing app with listen will cause 
+ * the unit tests to hang.
+ * 
+ * Although in jest, I think it can still apply to mocha
+ * Source: https://github.com/facebook/jest/issues/5783
+ * Article written about this case: https://www.albertgao.xyz/2017/05/24/how-to-test-expressjs-with-jest-and-supertest/
+ */
 const port = process.env.PORT || 8000;
 app.listen(port, () => {
   console.log(`Listening on port ${port}`)
