+++ backend/tests/spotify.spec.ts
@@ -1,0 +1,9 @@
+import { assert } from "chai"
+import request from "supertest"
+import app from "../src/app"
+
+describe("Spotify routes", () => {
+  it("/spotify/api", (done) => {
+    
+  });
+})
